Sanity checks on burden estimates

Helper function for sanity checks on burden estimate values. Checks whether any burden estimates are non-numeric, missing, or negative.

Usage

basic_burden_sanity(burden)

Arguments

burden

A <data.frame> of disease burden estimates. Must have at least a single column named "value" of numeric burden estimates.

Value

A character vector of messages generated by checks on burden estimates, with the length of the vector depending on how many checks fail.