vimcheck: Diagnostics for Vaccine Impact Modelling Consortium Burden and Impact Estimates

Run diagnostics on burden estimates submitted by modellers in the Vaccine Impact Modelling Consortium (VIMC) and impact estimates calculated by the VIMC Science & Policy team.

See also

Author

Maintainer: Pratik Gupte p.gupte24@imperial.ac.uk (ORCID)

Authors:

Katy Gaythorpe (ORCID)

Xiang Li (ORCID)

Zoë Gibney (ORCID)

Other contributors:

Imperial College of Science, Technology and Medicine [copyright holder, funder]